Page done, need comments
The navigation now works correctly. The animals are all called in correctly, and appear offscreen (except for the fox) until their animation calls them in. The animation for the fox makes him appear to jump, whereas with the bat and butterfly, the butterfly moves about halfway, stays for a while, then leaves. The bat goes up and down along the screen. The sprites all have correct styles and positioning. Need comments.